+++ lib/mailtrap.rb
@@ -10,9 +10,10 @@ require_relative 'mailtrap/contact_lists_api'
 require_relative 'mailtrap/contact_fields_api'
 require_relative 'mailtrap/contact_imports_api'
 require_relative 'mailtrap/suppressions_api'
+require_relative 'mailtrap/sending_domains_api'
 require_relative 'mailtrap/projects_api'
 require_relative 'mailtrap/inboxes_api'
-require_relative 'mailtrap/sending_domains_api'
+require_relative 'mailtrap/sandbox_messages_api'
 
 module Mailtrap
   # @!macro api_errors